Improve security of hash creation
With this patch I'm differentiating RequestTokenValues and SessionTokenValues
named tuples, to have the access_file hash stored in memory with the token itself

The advantage is we can use avoid to define the "salt" as global var and use
different salt for each hash creation.
Before of this page the same SALT was used for the whole DCVAuthenticator lifecycle.

Signed-off-by: Enrico Usai <usai@amazon.com>